Relocate classes
