+++ VWorkflows-Core/src/main/java/eu/mihosoft/vrl/workflow/impl/DefaultValueObject.java
@@ -31,8 +31,12 @@
  * authors and should not be interpreted as representing official policies, either expressed
  * or implied, of Michael Hoffer <info@michaelhoffer.de>.
  */
-package eu.mihosoft.vrl.workflow;
+package eu.mihosoft.vrl.workflow.impl;
 
+import eu.mihosoft.vrl.workflow.CompatibilityResult;
+import eu.mihosoft.vrl.workflow.VNode;
+import eu.mihosoft.vrl.workflow.ValueObject;
+import eu.mihosoft.vrl.workflow.VisualizationRequest;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
