+++ VWorkflows-Core/src/main/java/eu/mihosoft/vrl/workflow/impl/VFlowModelImpl.java
@@ -31,8 +31,20 @@
  * authors and should not be interpreted as representing official policies, either expressed
  * or implied, of Michael Hoffer <info@michaelhoffer.de>.
  */
-package eu.mihosoft.vrl.workflow;
+package eu.mihosoft.vrl.workflow.impl;
 
+import eu.mihosoft.vrl.workflow.Connection;
+import eu.mihosoft.vrl.workflow.ConnectionResult;
+import eu.mihosoft.vrl.workflow.Connections;
+import eu.mihosoft.vrl.workflow.Connector;
+import eu.mihosoft.vrl.workflow.FlowModel;
+import eu.mihosoft.vrl.workflow.IdGenerator;
+import eu.mihosoft.vrl.workflow.NodeLookup;
+import eu.mihosoft.vrl.workflow.ThruConnector;
+import eu.mihosoft.vrl.workflow.VFlowModel;
+import eu.mihosoft.vrl.workflow.VNode;
+import eu.mihosoft.vrl.workflow.ValueObject;
+import eu.mihosoft.vrl.workflow.VisualizationRequest;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.ObjectProperty;
@@ -53,7 +65,7 @@ import java.util.logging.Logger;
 /**
  * @author Michael Hoffer &lt;info@michaelhoffer.de&gt;
  */
-class VFlowModelImpl implements VFlowModel {
+public class VFlowModelImpl implements VFlowModel {
 
     private final VNodeImpl node;
     private final FlowModelImpl flow;
