+++ VWorkflows-Core/src/test/java/eu/mihosoft/vrl/workflow/ConnectorTest.java
@@ -33,6 +33,7 @@
  */
 package eu.mihosoft.vrl.workflow;
 
+import eu.mihosoft.vrl.workflow.util.FlowFactory;
 import org.junit.Assert;
 import org.junit.Test;
