+++ VWorkflows-Core/src/test/java/eu/mihosoft/vrl/workflow/SkinTest.java
@@ -37,6 +37,7 @@ import eu.mihosoft.vrl.workflow.skin.ConnectionSkin;
 import eu.mihosoft.vrl.workflow.skin.Skin;
 import eu.mihosoft.vrl.workflow.skin.SkinFactory;
 import eu.mihosoft.vrl.workflow.skin.VNodeSkin;
+import eu.mihosoft.vrl.workflow.util.FlowFactory;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import org.junit.Assert;
