+++ VWorkflows-Core/src/test/java/eu/mihosoft/vrl/workflow/VFlowMergeTest.java
@@ -38,6 +38,7 @@
  */
 package eu.mihosoft.vrl.workflow;
 
+import eu.mihosoft.vrl.workflow.util.FlowFactory;
 import org.junit.Ignore;
 
 /**
