+++ VWorkflows-Core/src/test/java/eu/mihosoft/vrl/workflow/VFlowTest.java
@@ -37,6 +37,7 @@
  */
 package eu.mihosoft.vrl.workflow;
 
+import eu.mihosoft.vrl.workflow.util.FlowFactory;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
