+++ VWorkflows-FX/src/main/java/eu/mihosoft/vrl/workflow/fx/ConnectorCircle.java
@@ -49,7 +49,7 @@ import javafx.scene.layout.Region;
  *
  * @author Michael Hoffer &lt;info@michaelhoffer.de&gt;
  */
-class ConnectorCircle extends Region implements ConnectorShape {
+public class ConnectorCircle extends Region implements ConnectorShape {
 
     private Connector connector;
     private final VFlow flow;
